core/requirements.txt
Django==1.10
djangorestframework==3.5.4
django-cors-middleware==1.3.1
django-rest-swagger==2.1.1
requests==2.13.0
python-dateutil==2.6.0
pika==0.10.0
django-cors-headers